update: split into controller and service
