+++ backend/src/routes/upload.routes.ts
@@ -1,32 +1,7 @@
 import { Router } from 'express';
-import supabase from '../database/supabase.init.js';
+import { createUploadUrls } from '../controllers/upload.controller.js';
 const router = Router();
 
-
-router.post('/upload-urls', async (req, res) => {
-	try {
-	  const { fileNames } = req.body
-	  if (!fileNames || !Array.isArray(fileNames)) {
-		return res.status(400).json({ error: 'Invalid fileNames array' })
-	  }
-
-	  const urls = []
-	  for (const name of fileNames) {
-		const { data, error } = await supabase.storage
-		  .from('user-photos')
-		  .createSignedUploadUrl(`users/${Date.now()}_${name}`)
-		if (error) {
-			console.error("Supabase signed URL error:", error)
-			return res.status(500).json({ error })
-		}
-		urls.push(data)
-	  }
-
-	  res.json({ urls })
-	} catch (err: any) {
-	  console.error(err)
-	  res.status(500).json({ error: 'Server error creating signed URLs' })
-	}
-  })
+router.post('/upload-urls', createUploadUrls)
 
 export default router;
